server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pydantic==2.4.2
pydantic-settings==2.0.3
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.7 
clerk-sdk-python>=2.0.0